# docs/api/swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "GHCAA.API, Version=1.0.0.0, Culture=neutral, PublicKeyToken=null",
    "version": "1.0"
  },
  "paths": {
...
    "/api/admin/members": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 1
            }
          },
          {
            "name": "pageSize",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 10
            }
          },
          {
            "name": "searchQuery",
            "in": "query",
            "schema": {
              "type": "string",
              "default": ""
            }
          },
          {
            "name": "statusFilter",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "categoryFilter",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "membershipTypeFilter",
            "in": "query",
            "schema": {
              "type": "string",
              "default": "all"
            }
          },
          {
            "name": "includeArchived",
            "in": "query",
            "schema": {
              "type": "boolean",
              "default": false
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMemberUpdateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMemberUpdateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/AdminMemberUpdateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/documents": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "patch": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "certificate": {
                    "type": "string",
                    "format": "binary"
                  },
                  "paymentProof": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "certificate": {
                  "style": "form"
                },
                "paymentProof": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/approve": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMemberDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMemberDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/reject": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMemberDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMemberDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/RejectMemberDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/members/{id}/restore": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/reactivate": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/photo": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "photo": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "photo": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/signature": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "signature": {
                    "type": "string",
                    "format": "binary"
                  }
                }
              },
              "encoding": {
                "signature": {
                  "style": "form"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/id-card": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/id-card/pdf": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/reset-password-admin": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/certificate": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/members/{id}/certificate/pdf": {
      "get": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/admin/contact-messages/{id}/read": {
      "post": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/admin/contact-messages/{id}": {
      "delete": {
        "tags": [
          "Admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AdminMemberUpdateDto": {
        "required": [
          "dateOfBirth",
          "email",
          "emergencyContactName",
          "emergencyContactPhone",
          "emergencyContactRelation",
          "fatherName",
          "fullName",
          "mobileNo",
          "motherName",
          "nid",
          "permanentAddress",
          "presentAddress"
        ],
        "type": "object",
        "properties": {
          "fullName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "fatherName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "motherName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "dateOfBirth": {
            "type": "string",
            "format": "date-time"
          },
          "nid": {
            "maxLength": 20,
            "minLength": 6,
            "type": "string"
          },
          "mobileNo": {
            "minLength": 1,
            "pattern": "^(\\+88)?01[3-9][\\s-]*\\d{8}$",
            "type": "string"
          },
          "email": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string",
            "format": "email"
          },
          "gender": {
            "$ref": "#/components/schemas/Gender"
          },
          "bloodGroup": {
            "$ref": "#/components/schemas/BloodGroup"
          },
          "presentAddress": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "permanentAddress": {
            "maxLength": 500,
            "minLength": 1,
            "type": "string"
          },
          "tShirtSize": {
            "maxLength": 10,
            "type": "string",
            "nullable": true
          },
          "emergencyContactName": {
            "maxLength": 200,
            "minLength": 1,
            "type": "string"
          },
          "emergencyContactRelation": {
            "maxLength": 100,
            "minLength": 1,
            "type": "string"
          },
          "emergencyContactPhone": {
            "minLength": 1,
            "pattern": "^(\\+88)?01[3-9][\\s-]*\\d{8}$",
            "type": "string"
          },
          "isMobilePublic": {
            "type": "boolean"
          },
          "isEmailPublic": {
            "type": "boolean"
          },
          "isAddressPublic": {
            "type": "boolean"
          },
          "isNIDPublic": {
            "type": "boolean"
          },
          "lastUpdateDate": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "notifyEventCreation": {
            "type": "boolean"
          },
          "notifyParticipationApproval": {
            "type": "boolean"
          },
          "notifyRegistrationUpdate": {
            "type": "boolean"
          },
          "notifyRelevantUpdates": {
            "type": "boolean"
          },
          "highestCertificate": {
            "type": "string",
            "nullable": true
          },
          "highestCertificateGroup": {
            "type": "string",
            "nullable": true
          },
          "highestCertificateSubject": {
            "type": "string",
            "nullable": true
          },
          "highestCertificatePassingYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "hscAdmissionYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ghcLastCertificate": {
            "type": "string",
            "nullable": true
          },
          "ghcLastCertificateGroup": {
            "type": "string",
            "nullable": true
          },
          "ghcLastCertificateSubject": {
            "type": "string",
            "nullable": true
          },
          "ghcLastCertificatePassingYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "ghcAdmissionYear": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "professionalSector": {
            "type": "string",
            "nullable": true
          },
          "designation": {
            "type": "string",
            "nullable": true
          },
          "organizationName": {
            "type": "string",
            "nullable": true
          },
          "location": {
            "type": "string",
            "nullable": true
          },
          "certificatePath": {
            "type": "string",
            "nullable": true
          },
          "paymentProofPath": {
            "type": "string",
            "nullable": true
          },
          "photoPath": {
            "type": "string",
            "nullable": true
          },
          "signaturePath": {
            "type": "string",
            "nullable": true
          },
          "academicHistory": {
            "minItems": 1,
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/AcademicRecordDto"
            },
            "nullable": true
          },
          "professionalHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ProfessionalRecordDto"
            },
            "nullable": true
          },
          "ecHistory": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ECHistoryDto"
            },
            "nullable": true
          },
          "status": {
            "$ref": "#/components/schemas/MembershipStatus"
          },
          "membershipType": {
            "$ref": "#/components/schemas/MembershipType"
          },
          "category": {
            "$ref": "#/components/schemas/MemberCategory"
          },
          "isVerified": {
            "type": "boolean"
          },
          "contributionPoints": {
            "type": "integer",
            "format": "int32"
          },
          "membershipNumber": {
            "type": "string",
            "nullable": true
          },
          "ecChangeReason": {
            "type": "string",
            "nullable": true
          },
          "membershipChangeReason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ApproveMemberDto": {
        "type": "object",
        "additionalProperties": false
      },
...
      "Gender": {
        "enum": [
          "None",
          "Male",
          "Female",
          "Other"
        ],
        "type": "string"
      },
...
    }
  }
}